Expression Calculator
Objective is to read a text file line by line containing arithmetic expressions, calculate them and print the original expression and the result onto a new text file. 
I have used two stack one for numbers and one for operators, but both of them are stacks of String because it makes it easy to compare string values at the end.
A method called "split()" is used to create an array of string values containing the original expression split into numbers and operators. Then we iterate over this array and perform our operation taking into consideration the parenthesis which have the highest precedence. It is possible that after the iteration over the array there are still some operations left to do, for this problem we run a while loop at the end until the operator stack is empty and finally print the original expression and result to the output file.